docs(gatsby/theme): refactor - utils - update/change parseAllFiles function
[skip ci]
